Move mux into separate file
