+++ internal/replacer/replacer.go
@@ -2,7 +2,6 @@
 package replacer
 
 import (
-	"errors"
 	"fmt"
 	"io"
 	"sync"
@@ -371,28 +370,4 @@ func mergeOverlaps(rs []subrange) []subrange {
 	rem := len(rs[j:]) // # of remaining ranges
 	copy(rs, rs[j:])
 	return rs[:rem]
-}
-
-// Mux contains multiple replacers
-type Mux []*Replacer
-
-// Flush flushes all replacers.
-func (mux Mux) Flush() error {
-	var errs []error
-	for _, r := range mux {
-		if err := r.Flush(); err != nil {
-			errs = append(errs, err)
-		}
-	}
-	if len(errs) != 0 {
-		return errors.Join(errs...)
-	}
-	return nil
-}
-
-// Reset resets all replacers with new needles (secrets).
-func (mux Mux) Reset(needles []string) {
-	for _, r := range mux {
-		r.Reset(needles)
-	}
 }
